+++ src/components/projectInfo/projectSummary/projectSumary.style.ts
@@ -2,142 +2,8 @@ import styled from 'styled-components';
 
 import { media } from '@/styles/media';
 
-const Container = styled.div`
-  color: #d6deec;
-  background: ${({ theme }) => theme.colors.primary.pri_900};
-  height: 100vh;
-  overflow: auto;
-  padding: 3% 6% 2% 6%;
-  display: flex;
-  flex-direction: column;
-  min-width: 410px;
-  min-height: 780px;
-  ${media.desktop`
-    .show {
-      opacity: 0;
-    }
-    margin-bottom: 30px;
-  `}
-  @media (max-width: 610px) {
-    .buttonShow {
-      opacity: 0;
-    }
-  }
-  @media (max-width: 1300px) {
-    overflow: auto;
-  }
-`;
-const Profile = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 0.813rem;
-  margin-bottom: 31.8px;
-  position: relative;
-`;
-
-const ProfileWrapper = styled.div`
-  min-width: 25.6px;
-  min-height: 25.6px;
-  width: 25.6px;
-  height: 25.6px;
-`;
-
-const ProfileName = styled.p`
-  font-family: Pretendard;
-  font-size: 19.2px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 150%;
-  letter-spacing: 0.384px;
-`;
-export const MemberBox = styled.div`
-  display: flex;
-  align-items: center;
-  width: 100%;
-  gap: 4.05%;
-  max-width: 100%;
-`;
-export const MemberName = styled.p`
-  margin: 0;
-  font-family: Pretendard;
-  font-size: 14.4px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 150%; /* 21.6px */
-  letter-spacing: 0.288px;
-  ${({ theme }) => theme.colors.white};
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  max-width: 80%;
-  @media (max-width: 1200px) {
-    max-width: 70%;
-  }
-`;
-export const ArrowWrapper = styled.div`
-  flex-shrink: 0;
-  flex: 1;
-`;
-export const Member = styled.div`
-  ${({ theme }) => theme.align.row_space_between};
-  width: 100%;
-  gap: 2.5%;
-  cursor: pointer;
-  margin-bottom: 20px;
-`;
-export const MemberContainer = styled.div`
-  overflow-y: scroll;
-  overflow-x: hidden;
-  width: 100%;
-  max-height: 78%;
-  ${media.desktop`
-    max-height: 80%;
-  `}
-`;
-const Box = styled.div<{ height: string }>`
-  height: ${(props) => props.height};
-  min-height: 100px;
-  background: var(--primary-pri_back_grad, linear-gradient(76deg, #0e2245 0%, #000714 100.13%));
-  border-radius: 8px;
-  padding: 16px 24px;
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-  border: 0.8px solid rgba(32, 75, 153, 0.2);
-`;
-
-const Left = styled.div`
-  width: 75%;
-  gap: 6%;
-  display: flex;
-  flex-direction: column;
-  justify-content: start;
-  ${media.desktop`
-    width: 100%;
-    height: 75%;
-  `}
-`;
-
-const Right = styled.div`
-  width: 22.06%;
-  overflow: hidden;
-  gap: 40px;
-  margin-top: 0;
-  display: flex;
-  flex-direction: column;
-  margin-left: 2.94112%;
-  min-width: 168px;
-  ${media.desktop`
-    width: 100%;
-    height: 50%;
-    margin: 40px 0 0 0;
-  `}
-`;
-
-const Title = styled.p`
+export const Title = styled.p`
   color: ${({ theme }) => theme.colors.primary.pri_50};
-  font-family: Pretendard;
   font-size: 17.6px;
   font-style: normal;
   font-weight: 500;
@@ -156,10 +22,18 @@ const Title = styled.p`
     margin-bottom: 10px;
   `}
 `;
-
-const TextLight = styled.p`
-  color: ${({ theme }) => theme.colors.white};
+export const TextBold = styled.p`
+  color: ${({ theme }) => theme.colors.primary.pri_50};
   font-family: Pretendard;
+  font-size: 22.4px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 150%;
+  letter-spacing: 0.448px;
+  margin-bottom: 8px;
+`;
+export const TextLight = styled.p`
+  color: ${({ theme }) => theme.colors.white};
   font-size: 11.2px;
   font-style: normal;
   font-weight: 500;
@@ -177,54 +51,12 @@ const TextLight = styled.p`
     white-space: pre;
   }
 `;
-
-const Text = styled.p`
-  max-width: 100%;
-  color: ${({ theme }) => theme.colors.white};
-  font-family: Pretendard;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 140%;
-  letter-spacing: 0.288px;
-  white-space: pre-line;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  scrollbar-width: none;
-  width: 85%;
-  word-wrap: break-word;
-  @media (max-width: 840px) {
-    width: 70%;
-  }
-  @media (max-width: 610px) {
-    width: 60%;
-    white-space: pre;
-  }
+export const Wrapper = styled.div`
+  position: absolute;
+  top: 16px;
+  right: 24px;
 `;
-
-const TextBold = styled.p`
-  color: ${({ theme }) => theme.colors.primary.pri_50};
-  font-family: Pretendard;
-  font-size: 22.4px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: 150%;
-  letter-spacing: 0.448px;
-  margin-bottom: 8px;
-`;
-
-const SemiBox = styled.div`
-  display: flex;
-  height: 70%;
-  margin-top: 3%;
-  ${media.desktop`
-    width: 100%;
-    height: 100%;
-    flex-wrap: wrap;
-  `}
-`;
-
-const InnerBox = styled.div`
+export const InnerBox = styled.div`
   box-sizing: border-box;
   background: rgba(217, 230, 255, 0.05);
   border-radius: 6.4px;
@@ -238,158 +70,6 @@ const InnerBox = styled.div`
     width: 100%;
     height: 100%;
   }
-`;
-
-const CharacterAddBox = styled.div`
-  ${({ theme }) => theme.align.column_center}
-  background: #d9e6ff1a;
-  border-radius: 6.4px;
-  height: inherit;
-  min-width: 112px;
-  padding: 12px;
-  cursor: pointer;
-  svg {
-    width: 50%;
-    height: 50%;
-  }
-  ${media.desktop`
-    height: 90%;
-  `}
-`;
-export const CharacterList = styled.div`
-  width: 85%;
-  height: 97%;
-  max-height: fit-content;
-  display: flex;
-  align-items: center;
-  gap: 16px;
-  overflow-x: scroll;
-  ${media.desktop`
-    height: 100%;
-  `}
-`;
-export const TooltipWrapper = styled.div<{ visible: boolean }>`
-  position: fixed;
-  z-index: 1000;
-  visibility: ${(props) => (props.visible ? 'visible' : 'hidden')};
-`;
-const CharacterBox = styled.div`
-  position: relative;
-  background: #007f7f;
-  border-radius: 6.4px;
-  width: 112px;
-  height: inherit;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  padding: 12px;
-  gap: 8px;
-  cursor: pointer;
-  ${media.desktop`
-    height: 90%;
-    flex-wrap: wrap;
-  `}
-`;
-const Character = styled.div`
-  display: flex;
-  height: 95%;
-  width: 100%;
-  align-items: center;
-  gap: 16px;
-  @media (max-width: 720px) {
-    overflow-x: scroll;
-  }
-  ${media.desktop`
-    align-items: flex-start;
-    height: 100%;
-  `}
-`;
-type TWrapperProps = {
-  top?: string;
-  bottom?: string;
-  right?: string;
-  left?: string;
-};
-
-export const Wrapper = styled.div<TWrapperProps>`
-  position: absolute;
-  top: ${(props) => props.top};
-  bottom: ${(props) => props.bottom};
-  left: ${(props) => props.left};
-  right: ${(props) => props.right};
-`;
-
-export const Input = styled.textarea`
-  width: 80%;
-  height: 100%;
-  color: ${({ theme }) => theme.colors.white};
-  font-family: Pretendard;
-  font-size: 11.2px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 150%; /* 21.6px */
-  letter-spacing: 0.288px;
-  padding: 8px;
-  border-radius: 6.4px;
-  background: rgba(217, 230, 255, 0.05);
-  border: none;
-  overflow: 'hidden';
-  resize: 'none';
-  @media (max-width: 840px) {
-    height: 100%;
-    width: 70%;
-  }
-  @media (max-width: 610px) {
-    width: 60%;
-  }
-  @media (max-width: 1400px) {
-    font-size: 12px;
-  }
-`;
-type TText = {
-  color?: string;
-};
-export const rowBox = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 8px;
-`;
-export const Medium18Text = styled.p<TText>`
-  color: ${(props) => props.color};
-  font-family: Pretendard;
-  font-size: 13px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 150%; /* 21.6px */
-  letter-spacing: 0.288px;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  max-width: 85%;
-`;
-export const Medium14Text = styled.p<TText>`
-  color: ${(props) => props.color};
-  font-family: Pretendard;
-  font-size: 11.2px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 150%; /* 16.8px */
-  letter-spacing: 0.224px;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  max-width: 85%;
-`;
-export const TableWrapper = styled.div`
-  display: flex;
-  flex: 1;
-  margin-left: 20px;
-  padding: 20px;
-  border-radius: 12.8px;
-  gap: 10px;
-  height: 283px;
-  overflow-x: scroll;
-  flex-direction: row;
 `;
 
 export const Table = styled.table`
@@ -460,65 +140,12 @@ export const TD = styled.td`
     width: 60%;
   }
 `;
-
-export const Wrap = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 3.5px;
-`;
 export const AccessRights = styled.div`
   display: flex;
   gap: 10px;
   align-items: center;
 `;
-export const ModalBox = styled.div`
-  gap: 20px;
-  margin: 20px 0;
-  display: flex;
-  flex-direction: column;
-`;
-export const PostBox = styled.div`
-  gap: 10px;
-  display: flex;
-  flex-direction: column;
-`;
-export const ModalText = styled.p`
-  ${({ theme }) => theme.text.medium_22};
-  color: ${({ theme }) => theme.colors.primary.pri_50};
-`;
-export const BtnWrapper = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: 10px;
-`;
-export const ProjectText = styled.p`
-  ${({ theme }) => theme.text.medium_24};
-  color: #d6deec;
-`;
-export const Position = styled.div`
-  display: flex;
-  justify-content: flex-end;
-`;
 export const Button = styled.div`
   flex: 0 0 auto;
   white-space: nowrap;
 `;
-export {
-  Box,
-  Character,
-  CharacterAddBox,
-  CharacterBox,
-  Container,
-  InnerBox,
-  Left,
-  Profile,
-  ProfileName,
-  ProfileWrapper,
-  Right,
-  SemiBox,
-  Text,
-  TextBold,
-  TextLight,
-  Title,
-};
